+++ test/unit/gemm/device/testbed_symm_universal.h
@@ -311,7 +311,7 @@ struct TestbedSymmUniversal {
     // Determine SMEM requirements and waive if not satisfied
     //
 
-    int smem_size = int(sizeof(typename Symm::SymmKernel::SharedStorage));
+    size_t smem_size = sizeof(typename Symm::SymmKernel::SharedStorage);
 
     cudaDeviceProp properties;
     int device_idx;
